2008-01-09  Stefan Huehner <stefan@huehner.org>
	* examples/gnu/classpath/examples/java2d/J2dBenchmark.java,
	* gnu/CORBA/OrbFocused.java,
	* gnu/CORBA/CDR/HeadlessInput.java,
	* gnu/java/awt/peer/gtk/CairoGraphics2D.java,
	* gnu/java/awt/peer/gtk/GtkFramePeer.java,
	* gnu/java/awt/peer/qt/QtMenuPeer.java,
	* gnu/java/nio/FileChannelImpl.java,
	* gnu/java/nio/SocketChannelSelectionKeyImpl.java,
	* gnu/java/security/Engine.java,
	* gnu/javax/imageio/jpeg/JPEGComponent.java,
	* gnu/javax/imageio/jpeg/JPEGDecoder.java,
	* gnu/javax/imageio/jpeg/JPEGFrame.java,
	* gnu/javax/print/PrinterDialog.java,
	* gnu/javax/security/auth/Password.java,
	* gnu/javax/swing/text/html/parser/support/Parser.java,
	* gnu/javax/swing/text/html/parser/support/parameterDefaulter.java,
	* gnu/xml/pipeline/DomConsumer.java,
	* java/awt/AWTKeyStroke.java,
	* java/awt/DefaultKeyboardFocusManager.java,
	* java/awt/GridBagLayout.java,
	* java/awt/dnd/DragGestureRecognizer.java,
	* java/awt/geom/Arc2D.java,
	* java/awt/geom/CubicCurve2D.java,
	* java/awt/geom/QuadCurve2D.java,
	* java/awt/im/InputContext.java,
	* java/awt/image/AffineTransformOp.java,
	* java/beans/Encoder.java,
	* java/io/ObjectInputStream.java,
	* java/io/ObjectStreamClass.java,
	* java/net/ResolverCache.java,
	* java/nio/charset/Charset.java,
	* java/security/SecureClassLoader.java,
	* java/text/MessageFormat.java,
	* java/util/jar/Manifest.java,
	* javax/accessibility/AccessibleRelationSet.java,
	* javax/imageio/ImageReader.java,
	* javax/imageio/metadata/IIOMetadataNode.java,
	* javax/print/attribute/standard/MediaSize.java,
	* javax/print/attribute/standard/PrinterStateReasons.java,
	* javax/swing/JTree.java,
	* javax/swing/table/DefaultTableColumnModel.java,
	* javax/swing/text/html/MinimalHTMLWriter.java,
	* javax/swing/text/html/parser/DTD.java,
	* javax/swing/tree/DefaultMutableTreeNode.java,
	* javax/swing/undo/UndoManager.java,
	* org/omg/CosNaming/NamingContextExtPOA.java,
	* org/omg/CosNaming/NamingContextPOA.java,
	* org/omg/CosNaming/_NamingContextExtImplBase.java,
	* org/omg/CosNaming/_NamingContextImplBase.java,
	* tools/gnu/classpath/tools/appletviewer/CommonAppletStub.java:
	Remove unneeded casts

+++ gnu/java/awt/peer/gtk/CairoGraphics2D.java
@@ -2114,10 +2114,10 @@ public abstract class CairoGraphics2D extends Graphics2D
   private static Rectangle computeIntersection(int x, int y, int w, int h,
                                                Rectangle rect)
   {
-    int x2 = (int) rect.x;
-    int y2 = (int) rect.y;
-    int w2 = (int) rect.width;
-    int h2 = (int) rect.height;
+    int x2 = rect.x;
+    int y2 = rect.y;
+    int w2 = rect.width;
+    int h2 = rect.height;
 
     int dx = (x > x2) ? x : x2;
     int dy = (y > y2) ? y : y2;
